requirements.txt
Django==4.2.3
celery==5.2.7
redis==4.3.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.7
psycopg==3.1.10
pandas==2.0.3
numpy==1.23.5